% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Creation of a boxplot}
\usage{
boxplot(
  dataset,
  xvar,
  yvar,
  fill,
  notch,
  widthnotch,
  graphorientation,
  legend,
  xlabel,
  ylabel,
  title
)
}
\arguments{
\item{dataset}{what dataset you want to be used in the}

\item{xvar}{is the variable that x will display.}

\item{yvar}{is the variable that y will display.
This  geom_boxplot specifies a boxplot will be created.}

\item{fill}{is the color you want to fill the boxplots with or the factor you want to use to separate the data.}

\item{notch}{is a TRUE or FALSE input that decides if a notch will be in the boxplot.}

\item{widthnotch}{is the size of the notch.}

\item{graphorientation}{is either x (which will rotate the graph so it is vertical),
or y (which will rotate the graph so that it is horizontal).}

\item{legend}{is a true or false value that will determine if the legend is there or not.
labs creates the label}

\item{xlabel}{determines what the x axis is labeled.}

\item{ylabel}{determines what the y axis is labeled.}

\item{title}{specifies what the title of the graph is.}
}
\description{
`boxplot` returns a created boxplot based on the dataset and characteristics
inputed into the function, as well as labels for title and axis.
}
\details{
This function has mamy arguments that can be added or removed to make
the boxplot more flexible to the visual wanted and more representative of
the data.
}
