% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C18factors.R
\docType{data}
\name{C18factors}
\alias{C18factors}
\title{Various factors that affect the metabolites in a person}
\format{
A data frame with 546 rows and 10 variables:
}
\usage{
C18factors
}
\description{
A dataset that contains various factors measured about the volunteers,
including attributes like gender and diagnosis. This is comparing those
with Crohn's Disease, Ulcerative Colitis and no diagnosis.
}
\keyword{datasets}
