% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C18named.R
\docType{data}
\name{C18named}
\alias{C18named}
\title{Samples of Many Metabolites}
\format{
A data frame with 183 rows and 547 variables:
}
\usage{
C18named
}
\description{
A dataset that shows the various metabolites measured in a study. The
samples match those of the C18factor dataset as well.
}
\keyword{datasets}
