% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Creation of a heatmap.}
\usage{
heatmap(
  dataset,
  rowdendrogram,
  coldendrogram,
  heatmapcolor,
  ycolor,
  xcolor,
  title,
  xlabel,
  ylabel
)
}
\arguments{
\item{dataset}{what dataset you want to be used in the creation of the heatmap.}

\item{rowdendrogram}{informs the program whether or not a dendrogram should be included on the y axis.}

\item{coldendrogram}{informs the program whether or not a dendrogram should be included on the y axis.}

\item{heatmapcolor}{is what color the heatmap with be created in shades of.}

\item{ycolor}{is what column (column should have colored values) the y axis should be colored by.}

\item{xcolor}{is what column (column should have colored values) the x axis should be colored by.}

\item{title}{is the title of the graph}

\item{xlabel}{is what the x axis label will be}

\item{ylabel}{is what the y axis label wil be.}
}
\description{
`heat_map` returns a created heatmap based on the dataset and characteristics
you enter into the function, as well as labels.
}
\details{
This function has several arguments that can be added or removed to make
the heatmap more flexible to the visual wanted.
}
