% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ec.R
\name{heatmap_ec}
\alias{heatmap_ec}
\title{Creation of a heatmap.}
\usage{
heatmap_ec(
  dataset,
  rowdendrogram = NULL,
  coldendrogram = NULL,
  heatmapcolor = brewer.pal(9, "Blues"),
  ycolor,
  xcolor,
  title = NA,
  xlabel = NA,
  ylabel = NA
)
}
\arguments{
\item{dataset}{what dataset you want to be used in the creation of the heatmap.}

\item{rowdendrogram}{informs the program whether or not a dendrogram should be included on the y axis. By default it will create the dendrogram. Setting it equal to NA will get rid of the dendrogram.}

\item{coldendrogram}{informs the program whether or not a dendrogram should be included on the y axis. By default it will create the dendrogram. Setting it equal to NA will get rid of the dendrogram.}

\item{heatmapcolor}{is what color the heatmap with be created in shades of.}

\item{ycolor}{is what column (column should have colored values) the y axis should be colored by.}

\item{xcolor}{is what column (column should have colored values) the x axis should be colored by.}

\item{title}{is the title of the graph}

\item{xlabel}{is what the x axis label will be}

\item{ylabel}{is what the y axis label wil be.}
}
\description{
`heat_map_ec` returns a created heatmap based on the dataset and characteristics
you enter into the function, as well as labels.
}
\details{
This function has several arguments that can be added or removed to make
the heatmap more flexible to the visual wanted.
}
