% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot_ec.R
\name{lineplot_ec}
\alias{lineplot_ec}
\title{Creation of a line graph}
\usage{
lineplot_ec(
  dataset,
  xvar = NULL,
  yvar = NULL,
  color = NULL,
  linesize = NULL,
  legend = NA,
  poslegend = NULL,
  xlabel = NULL,
  ylabel = NULL,
  title = NULL
)
}
\arguments{
\item{dataset}{what dataset you want to be used in the creation of the rest of the line graph.}

\item{xvar}{is the variable that x will display.}

\item{yvar}{is the variable that y will display.}

\item{color}{is what color the graph will be, it can also be the separating factor to make a multiple line line graph.}

\item{linesize}{is the width of the lines.}

\item{legend}{is a TRUE or FALSE value on if a legend should be included.
theme adds another specification we can add.}

\item{poslegend}{is where on the graph the legend should be. Acceptable arguments include: "left", "right", "bottom", "top".
labs creates the label.}

\item{xlabel}{determines what the x axis is labeled.}

\item{ylabel}{determines what the y axis is labeled.}

\item{title}{specifies what the title of the graph is.}
}
\description{
`lineplot_ec` returns a created line graph based on the dataset and
various specifications inputed into the function, as well as labels
for title and axis.
}
\details{
This function has several arguments that can be used or ignored to make
the boxplot more flexible to the visual wanted and more representative of
the data.

The ggplot sets up the creation of a plot.
}
