% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaplot_ec.R
\name{pcaplot_ec}
\alias{pcaplot_ec}
\title{Creation of a PCA plot}
\usage{
pcaplot_ec(
  dataset,
  ellipse = FALSE,
  groupedby = NULL,
  axes = FALSE,
  adjustnamevar = 1,
  namevarsize = 5,
  xlabel = NULL,
  ylabel = NULL,
  title = NULL
)
}
\arguments{
\item{dataset}{what dataset you want to be used in the
creation of the rest of the PCA plot.}

\item{ellipse}{is a TRUE or FALSE input that determines if normal ellipsis are used.}

\item{groupedby}{is what the ellipsis are grouped into.}

\item{axes}{is a TRUE or FALSE input deciding if arrows for variables are created.}

\item{adjustnamevar}{is how far the name / label is from the arrow.}

\item{namevarsize}{is how large the labels are for the arrows
labs creates the label}

\item{xlabel}{determines what the x axis is labeled.}

\item{ylabel}{determines what the y axis is labeled.}

\item{title}{specifies what the title of the graph is.}
}
\description{
`pcaplot_ec` returns a created PCA plot based on the dataset and
various specifications inserted into the function, as well as labels
for title and axis.
}
\details{
This function has several arguments that can be implemented to make
the PCA plot more flexible to the visual wanted and more representative of
the data.

The ggplot sets up the creation of a plot.
}
